Reduce image size to 1.5G
/opt/nwnserver/data was huge (1.7G) and contained standard NWN resources, the
duplicates of which can be found in the data directory of any NWN client install
path. I cannot see a reason why anyone would want to use the image without a
client already installed, and so the folder was cleaned. Run with -v, as
explained in github.com/jakkn/nwn-devbase/DOCKERGUIDE.md.

/opt/nwnserver/nvm was about 70M, and I cannot see a reason why anyone would
want to use them. If you want it, mount with -v.

apt-cache cleaned for good measure.

Total image size reduced from 5.7G to 1.5G.